+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- list_folders method for GcsBucket - [#121](https://github.com/PrefectHQ/prefect-gcp/pull/121)
+- `vpc_connector_name` field to `CloudRunJob` - [#123](https://github.com/PrefectHQ/prefect-gcp/pull/123)
+- `list_folders` method for GcsBucket - [#121](https://github.com/PrefectHQ/prefect-gcp/pull/121)
+
 ### Changed
 
 ### Deprecated